Refactor modules, allow backwards compat with asyncio
